calvin_models/requirements.txt
cmake==3.18.4
wheel
numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib
opencv-python
omegaconf
plotly
pytorch-lightning==1.8.6
lightning_lite
torch==1.13.1
torchvision
MulticoreTSNE
gitpython
scipy
sentence-transformers
setuptools==57.5.0
gym
moviepy
tqdm
termcolor
wandb
pyhash
